+++ src/main/java/chess/model/dto/BoardDTO.java
@@ -1,4 +1,4 @@
-package chess.model;
+package chess.model.dto;
 
 import java.util.List;
 import java.util.Objects;
