+++ src/main/java/chess/model/piece/Piece.java
@@ -1,8 +1,7 @@
 package chess.model.piece;
 
-import chess.model.Coordinate;
-import chess.model.Route;
-import chess.model.Vector;
+import chess.model.board.Coordinate;
+import chess.model.vector.Vector;
 
 import java.util.List;
