+++ src/main/java/chess/model/vector/Direction.java
@@ -1,4 +1,6 @@
-package chess.model;
+package chess.model.vector;
+
+import chess.model.board.Coordinate;
 
 import java.util.List;
