+++ src/test/java/chess/model/BoardTest.java
@@ -1,16 +1,20 @@
 package chess.model;
 
+import chess.model.board.Board;
+import chess.model.board.Coordinate;
+import chess.model.board.Tile;
+import chess.model.boardcreatestrategy.NewGameCreateStrategy;
 import chess.model.piece.King;
 import chess.model.piece.Knight;
 import chess.model.piece.Pawn;
 import chess.model.piece.Queen;
 import chess.model.piece.Empty;
+import chess.model.vector.Vector;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
