+++ src/test/java/chess/model/ChessGameTest.java
@@ -1,5 +1,6 @@
 package chess.model;
 
+import chess.model.boardcreatestrategy.NewGameCreateStrategy;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
