+++ src/test/java/chess/model/CoordinateTest.java
@@ -1,5 +1,6 @@
 package chess.model;
 
+import chess.model.board.Coordinate;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
