+++ src/test/java/chess/model/DirectionTest.java
@@ -1,5 +1,7 @@
 package chess.model;
 
+import chess.model.board.Coordinate;
+import chess.model.vector.Direction;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
