+++ src/test/java/chess/model/VectorTest.java
@@ -1,5 +1,9 @@
 package chess.model;
 
+import chess.model.board.Coordinate;
+import chess.model.vector.Direction;
+import chess.model.vector.Magnitude;
+import chess.model.vector.Vector;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
